Fix compile warnings

+++ lib/ithimetrics.cpp
@@ -559,8 +559,8 @@ bool ithimetrics::ParseMetricFileName(const char * name, int * metric_id, int * 
          * '.' after day
          */
         char sep_expect[4] = { '-', '-', '-', '.' };
-        int sep_found[4] = { 0, 0, 0, 0 };
-        int current = (int)char_after_sep_index;
+        size_t sep_found[4] = { 0, 0, 0, 0 };
+        size_t current = char_after_sep_index;
         int nb_found = 0;
         int current_val = 0;
